feat: improve notice styling, include 'restart' button in authorization notice, extract restart logic into OverlayNotice
